requirements.txt
google-genai==1.31.0
python-dotenv==1.1.1
Flask==3.1.2
pypdf==6.0.0